+++ {{cookiecutter.project_name}}/frontend/components/Hero/Hero.js
@@ -1,15 +1,10 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import i18n from '../../i18n';
 import s from './Hero.module.css';
 
-const Hero = ({ title }) => (
+const Hero = ({ title = '' }) => (
     <div className={s.Container}>
-        <img
-            className={s.Logo}
-            src="/img/logo.svg"
-            alt="Logo"
-        />
+        <img className={s.Logo} src="/img/logo.svg" alt="Logo" />
         <h1 className={s.Title}>
             <img
                 src={'/img/white_circle.png'}
@@ -21,13 +16,5 @@ const Hero = ({ title }) => (
         <p className={s.Preamble}>{i18n.t('helloWorld')}</p>
     </div>
 );
-
-Hero.propTypes = {
-    title: PropTypes.string,
-};
-
-Hero.defaultProps = {
-    title: '',
-};
 
 export default Hero;