+++ {{cookiecutter.project_name}}/frontend/components/RawHtml/RawHtml.js
@@ -1,17 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import s from './RawHtml.module.css';
 
-const RawHtml = ({ html }) => (
+const RawHtml = ({ html = '' }) => (
     <div className={s.Container} dangerouslySetInnerHTML={{ __html: html }} />
 );
-
-RawHtml.propTypes = {
-    html: PropTypes.string,
-};
-
-RawHtml.defaultProps = {
-    html: '',
-};
 
 export default RawHtml;